refactor: Moves ObjectCloner, implementation and facade to cross-cutting-concerns.
The interface is put in the domain as there are usecases in the domain to make a copy of an object there as well as in the other layers.
The implementation is put in the infrastructure as it could be using external frameworks/libs.
The facade is currently put in the application layer since it is used in the Tv2ActionFactoryProvider - this is the wrong place and it should be remodelled. It is noted and will be corrected later on.